fix FLASHViewer to match merging with FLASHTaggerViewer
